requirements.txt
# Dependencias principales
sqlalchemy==2.0.23
anthropic==0.5.0
openai==1.3.0
pandas==2.1.1
python-dotenv==1.0.0

# Utilidades
tqdm==4.66.1
rich==13.6.0
colorama==0.4.6

# API y Web
fastapi==0.104.1
uvicorn==0.24.0
flask==2.3.3
jinja2==3.1.2

# CLI
click==8.1.7
prompt_toolkit==3.0.39

# Testing
pytest==7.4.3
pymysql==1.1.0
